+++ src/main/java/com/kruemblegard/client/render/model/CephalariZombieModel.java
@@ -1,0 +1,25 @@
+package com.kruemblegard.client.render.model;
+
+import com.kruemblegard.Kruemblegard;
+import com.kruemblegard.entity.CephalariZombieEntity;
+
+import net.minecraft.resources.ResourceLocation;
+
+import software.bernie.geckolib.model.GeoModel;
+
+public class CephalariZombieModel extends GeoModel<CephalariZombieEntity> {
+    @Override
+    public ResourceLocation getModelResource(CephalariZombieEntity animatable) {
+        return new ResourceLocation(Kruemblegard.MOD_ID, "geo/cephalari_zombie.geo.json");
+    }
+
+    @Override
+    public ResourceLocation getTextureResource(CephalariZombieEntity animatable) {
+        return new ResourceLocation(Kruemblegard.MOD_ID, "textures/entity/cephalari_zombie.png");
+    }
+
+    @Override
+    public ResourceLocation getAnimationResource(CephalariZombieEntity animatable) {
+        return new ResourceLocation(Kruemblegard.MOD_ID, "animations/cephalari_zombie.animation.json");
+    }
+}